Add mkMap property to the module system
This allows specifying configuration that should be merged in with some
or all elements of a container option (e.g. listOf, attrsOf) without
needing to know whether those elements are actually defined elsewhere.

For example, this will default the fsType of all of your fileSystems
except /boot to btrfs (and leaves /boot alone):

    fileSystems = mkMap (name: if name != "/boot"
        then [ { fsType = mkDefault "btrfs"; } ]
        else []
    );

mkMap takes a function which takes an index of an element and returns a
list of values to be merged in with other values defined for that
element. For sets, the index is just the attribute name. For lists, the
mkMap function is passed two arguments: The index of the relevant
definition in the list of all definitions for that option, and the index
of the element within that definition.